model/requirements.txt
Flask==3.1.2
flask_cors==6.0.1
gTTS==2.5.4
openai_whisper==20250625
protobuf==6.33.0
PyJWT==2.10.1
python-dotenv==1.1.1
Requests==2.32.5
supabase==2.22.0
